refactor: experience items service into function
