+++ src/app/resume-page/experience-section/get-experience-items.ts
@@ -1,0 +1,25 @@
+import { inject, InjectionToken } from '@angular/core'
+import resume from '../../../../assets/resume.json'
+import { ADAPT_JSON_RESUME_WORK } from './adapt-json-resume-work'
+import { ExperienceItem } from './experience-item/experience-item'
+
+export type GetExperienceItems = () => ReadonlyArray<ExperienceItem>
+export const GET_EXPERIENCE_ITEMS = new InjectionToken<GetExperienceItems>(
+  isDevMode ? 'GetExperienceItems' : 'GEI',
+  {
+    factory: () => {
+      const works = inject(JSON_RESUME_WORKS)
+      const adapter = inject(ADAPT_JSON_RESUME_WORK)
+      return () => works.map((work) => adapter(work))
+    },
+  },
+)
+
+export const JSON_RESUME_WORKS = new InjectionToken<JsonResumeWorks>(
+  isDevMode ? 'JSON Resume works' : 'JRWs',
+  {
+    factory: () => resume.work,
+  },
+)
+
+export type JsonResumeWorks = typeof resume.work
